убран альтернативный функционал, вынесен интерфейс и UnsupportedTypeException, много косметических исправлений.
