+++ hw08-Json/src/main/java/ru/otus/homework/diyjson/DiyJsonWriter.java
@@ -1,0 +1,7 @@
+package ru.otus.homework.diyjson;
+
+public interface DiyJsonWriter {
+
+    String toJson(Object obj) throws UnsupportedTypeException;
+}
+
