+++ hw08-Json/src/main/java/ru/otus/homework/diyjson/UnsupportedTypeException.java
@@ -1,0 +1,11 @@
+package ru.otus.homework.diyjson;
+
+public class UnsupportedTypeException extends Exception {
+
+    public UnsupportedTypeException() {
+    }
+
+    public UnsupportedTypeException(String s) {
+        super(s);
+    }
+}
