--- hw08-Json/src/main/java/ru/otus/homework/diyjson/UnsupportedTypeExeption.java
@@ -1,4 +1,0 @@
-package ru.otus.homework.diyjson;
-
-public class UnsupportedTypeExeption {
-}
