+++ hw08-Json/src/main/java/ru/otus/homework/diyjson/forTest/Parent.java
@@ -1,7 +1,9 @@
 package ru.otus.homework.diyjson.forTest;
 
 public class Parent {
-    private String parentPrivate = "parentPrivateVal";
-    protected String parentProtected = "parentProtectedVal";
     public String parentPublic = "parentPublicVal";
+    protected String parentProtected = "parentProtectedVal";
+    private String parentPrivate = "parentPrivateVal";
+
+
 }
